Function buf_redux::copy_buf

pub fn copy_buf<B: BufRead, W: Write>(b: &mut B, w: &mut W) -> Result<u64>

Copy data between a BufRead and a Write without an intermediate buffer.

Retries on interrupts. Returns the total bytes copied or the first error; even if an error is returned some bytes may still have been copied.