Struct buf_redux::LineWriter

pub struct LineWriter<W: Write>(_);

A drop-in replacement for std::io::LineWriter with more functionality.

Methods

impl<W: Write> LineWriter<W>

pub fn new(inner: W) -> Self

Wrap inner with the default buffer capacity.

pub fn with_capacity(capacity: usize, inner: W) -> Self

Wrap inner with the given buffer capacity.

pub fn get_ref(&self) -> &W

Get a reference to the inner writer.

pub fn get_mut(&mut self) -> &mut W

Get a mutable reference to the inner writer.

Note

If the buffer has not been flushed, writing directly to the inner type will cause data inconsistency.

pub fn capacity(&self) -> usize

Get the capacty of the inner buffer.

pub fn buffered(&self) -> usize

Get the number of bytes currently in the buffer.

pub fn grow(&mut self, additional: usize)

Grow the internal buffer by at least additional bytes. May not be quite exact due to implementation details of the buffer's allocator.

Note

This should not be called frequently as each call will incur a reallocation.

pub fn into_inner(self) -> Result<W, IntoInnerError<Self>>

Flush the buffer and unwrap, returning the inner writer on success, or a type wrapping self plus the error otherwise.

pub fn into_inner_with_err(self) -> (W, Option<Error>)

Flush the buffer and unwrap, returning the inner writer and any error encountered during flushing.

pub fn into_inner_with_buf(self) -> (W, Vec<u8>)

Consume self and return both the underlying writer and the buffer, with the data moved to the beginning and the length truncated to contain only valid data.