Struct buf_redux::BufReader

pub struct BufReader<R, Rs = DefaultReadStrategy, Ms = DefaultMoveStrategy> { /* fields omitted */ }

A drop-in replacement for std::io::BufReader with more functionality.

Original method names/signatures and implemented traits are left untouched, making replacement as simple as swapping the import of the type.

Methods

impl<R> BufReader<R, DefaultReadStrategy, DefaultMoveStrategy>

pub fn new(inner: R) -> Self

Create a new BufReader wrapping inner, with a buffer of a default capacity and the default strategies.

pub fn with_capacity(cap: usize, inner: R) -> Self

Create a new BufReader wrapping inner with a capacity of at least cap bytes and the default strategies.

The actual capacity of the buffer may vary based on implementation details of the buffer's allocator.

impl<R, Rs: ReadStrategy, Ms: MoveStrategy> BufReader<R, Rs, Ms>

pub fn with_strategies(inner: R, rs: Rs, ms: Ms) -> Self

Create a new BufReader wrapping inner, with a default buffer capacity and with the given ReadStrategy and MoveStrategy.

pub fn with_cap_and_strategies(inner: R, cap: usize, rs: Rs, ms: Ms) -> Self

Create a new BufReader wrapping inner, with a buffer capacity of at least cap bytes and the given ReadStrategy and MoveStrategy.

The actual capacity of the buffer may vary based on implementation details of the buffer's allocator.

pub fn move_strategy<Ms_: MoveStrategy>(self, ms: Ms_) -> BufReader<R, Rs, Ms_>

Apply a new MoveStrategy to this BufReader, returning the transformed type.

pub fn read_strategy<Rs_: ReadStrategy>(self, rs: Rs_) -> BufReader<R, Rs_, Ms>

Apply a new ReadStrategy to this BufReader, returning the transformed type.

pub fn move_strategy_mut(&mut self) -> &mut Ms

Accessor for updating the MoveStrategy in-place.

If you want to change the type, use .move_strategy().

pub fn read_strategy_mut(&mut self) -> &mut Rs

Accessor for updating the ReadStrategy in-place.

If you want to change the type, use .read_strategy().

pub fn make_room(&mut self)

Move data to the start of the buffer, making room at the end for more reading.

pub fn grow(&mut self, additional: usize)

Grow the internal buffer by at least additional bytes. May not be quite exact due to implementation details of the buffer's allocator.

Note

This should not be called frequently as each call will incur a reallocation.

pub fn get_buf(&self) -> &[u8]

Get the section of the buffer containing valid data; may be empty.

Call .consume() to remove bytes from the beginning of this section.

pub fn available(&self) -> usize

Get the current number of bytes available in the buffer.

pub fn capacity(&self) -> usize

Get the total buffer capacity.

pub fn get_ref(&self) -> &R

Get an immutable reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Get a mutable reference to the underlying reader.

Note

Reading directly from the underlying reader is not recommended, as some data has likely already been moved into the buffer.

pub fn into_inner(self) -> R

Consume self and return the inner reader only.

pub fn into_inner_with_buf(self) -> (R, Vec<u8>)

Consume self and return both the underlying reader and the buffer, with the data moved to the beginning and the length truncated to contain only valid data.

See also: BufReader::unbuffer()

pub fn unbuffer(self) -> Unbuffer<R>

Consume self and return an adapter which implements Read and will empty the buffer before reading directly from the underlying reader.

impl<R: Read, Rs: ReadStrategy, Ms: MoveStrategy> BufReader<R, Rs, Ms>

pub fn read_into_buf(&mut self) -> Result<usize>

Unconditionally perform a read into the buffer, calling .make_room() if appropriate or necessary, as determined by the implementation.

If the read was successful, returns the number of bytes read.

impl<R: Read, Rs, Ms> BufReader<R, Rs, Ms>

pub fn boxed<'a>(self) -> BufReader<Box<Read + 'a>, Rs, Ms> where
R: 'a,

Box the inner reader without losing data.